+++ src/Helper.php
@@ -283,4 +283,16 @@ class Helper {
 		// return the requested filesystem object.
 		return $wp_filesystem;
 	}
+
+	/**
+	 * Return whether this WordPress runs in development mode (available since WordPress 6.3).
+	 *
+	 * @return bool
+	 */
+	public static function is_development_mode(): bool {
+		return (
+			       function_exists( 'wp_is_development_mode' ) && false !== wp_is_development_mode( 'plugin' )
+		       )
+		       || ! function_exists( 'wp_is_development_mode' );
+	}
 }